Simplify test for --tmpdir in package deploy
